tools: smmute: Add reproducer for TLB issue
There is a bug in the model (will be fixed soon), that prevents the device
from accessing a valid mapping if that same address previously caused a
Invalid Request PRI reponse. The model would omit to drop the channel
between device and DRAM for that transaction, and instead redirect
subsequent identical ones to a sink.

Add a test case to easily reproduce it (and test for regressions):

	- open device
	- map X
	- transaction on X
	  -> succeeds
	- unmap X
	- transaction in X
	  -> fails as expected -- Invalid Request
	- map X
	- transaction on X
	  -> fails!
	- unmap X
	- close device, retry.

Signed-off-by: Jean-Philippe Brucker <jean-philippe.brucker@arm.com>